requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
altgraph==0.17
appdirs==1.4.4
asgiref==3.4.1
async-timeout==4.0.2
attrs==21.2.0
auto-py-to-exe==2.11.0
autocrop==1.1.1
autopep8==1.5.7
backports.entry-points-selectable==1.1.1
base58==2.1.0
beautifulsoup4==4.9.3
bitarray==1.2.2
bottle==0.12.19
bottle-websocket==0.2.9
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
charset-normalizer==2.0.11
click==7.1.2
colorama==0.4.4
cryptography==3.4.7
csv342==1.0.1
cycler==0.10.0
cytoolz==0.11.0
distlib==0.3.1
Django==3.2.5
django-bootstrap-v5==1.0.4
django-classy-tags==2.0.0
django-cms==3.9.0
django-formtools==2.3
django-sekizai==2.0.0
django-treebeard==4.5.1
djangocms-admin-style==2.0.2
djangocms-animate-on-scroll==1.0.1
docutils==0.18.1
Eel==0.12.4
eth-abi==2.1.1
eth-account==0.5.4
eth-hash==0.3.1
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.10.0
fastapi==0.65.1
filelock==3.4.0
Flask==2.0.2
frozenlist==1.3.0
future==0.18.2
gevent==21.8.0
gevent-websocket==0.10.1
greenlet==1.1.2
gunicorn==20.1.0
h11==0.12.0
hexbytes==0.2.1
idna==2.10
importlib==1.0.4
ipfshttpclient==0.7.0a1
itsdangerous==2.0.1
Jinja2==3.0.3
jsonschema==3.2.0
keyboard==0.13.5
Kivy==2.0.0
kivy-deps.angle==0.3.0
kivy-deps.glew==0.3.0
kivy-deps.gstreamer==0.3.2
kivy-deps.sdl2==0.3.1
Kivy-examples==2.0.0
Kivy-Garden==0.1.4
kiwisolver==1.3.1
line-bot-sdk==2.1.0
lru-dict==1.1.7
MacTmp==0.0.6
MarkupSafe==2.0.1
matplotlib==3.4.2
MouseInfo==0.1.3
multiaddr==0.0.9
multidict==6.0.2
mysqlclient==2.0.3
netaddr==0.8.0
numpy @ file:///C:/Users/NADOOD/Desktop/sakc/packet/numpy-1.21.0-cp39-cp39-win_amd64.whl
opencv-python==4.5.2.52
pandas @ file:///C:/Users/NADOOD/Desktop/sakc/packet/pandas-1.3.0-cp39-cp39-win_amd64.whl
parsimonious==0.8.1
pefile==2021.5.24
Pillow==8.1.2
platformdirs==2.4.0
progress==1.5
progressbar2==3.53.1
protobuf==3.16.0
PyAutoGUI==0.9.53
pycodestyle==2.7.0
pycparser==2.20
pycryptodome==3.10.1
pydantic==1.8.2
PyGetWindow==0.0.9
Pygments==2.10.0
pyinstaller==4.7
pyinstaller-hooks-contrib==2021.2
pyLoRa==0.3.1
pyModbusTCP==0.1.10
PyMsgBox==1.0.9
PyMySQL==1.0.2
pynput==1.7.3
pyodbc @ file:///C:/Users/NADOOD/Desktop/sakc/packet/pyodbc-4.0.31-cp39-cp39-win_amd64.whl
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyperclip==1.8.2
pypiwin32==223
pypng==0.0.20
PyQRCode==1.2.1
PyQt5==5.15.3
PyQt5-Qt==5.15.2
PyQt5-sip==12.8.1
PyRect==0.1.4
pyrsistent==0.17.3
PyScreeze==0.1.27
python-dateutil @ file:///C:/Users/NADOOD/Desktop/sakc/packet/python_dateutil-2.8.2-py2.py3-none-any.whl
python-utils==2.5.6
PyTweening==1.0.3
pytz==2021.1
pyuic5-tool==0.0.1
pywin32==300
pywin32-ctypes==0.2.0
qrcode==6.1
requests==2.25.1
rlp==2.0.1
scipy==1.7.0
seaborn==0.11.1
six @ file:///C:/Users/NADOOD/Desktop/sakc/packet/six-1.16.0-py2.py3-none-any.whl
soupsieve==2.2.1
sqlparse==0.4.1
starlette==0.14.2
toml==0.10.2
toolz==0.11.1
tqdm==4.61.1
typing-extensions==3.10.0.0
urllib3==1.26.4
uvicorn==0.13.4
varint==1.0.2
virtualenv==20.10.0
virtualenvwrapper-win==1.2.6
web3==5.19.0
websockets==8.1
Werkzeug==2.0.3
whichcraft==0.6.1
yarl==1.7.2
zope.event==4.5.0
zope.interface==5.4.0
